requirements.txt
-i https://pypi.org/simple
argparse==1.4.0
asgiref==3.2.10; python_version >= '3.5'
certifi==2020.6.20
chardet==3.0.4
dateutils==0.6.11
django-cors-headers==3.1.1
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-rest-framework-braces==0.3.4
django==3.1.2
djangorestframework==3.12.1; python_version >= '3.5'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.1
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
